pyproject.toml
[tool.poetry]
name = "file-pipe"
version = "0.1.0"
description = "TBD"
authors = ["nawazsk"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.116.1"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
pydantic = "^2.11.7"
pydantic-settings = "^2.10.1"
jwt = "^1.4.0"
python-jose = "^3.5.0"
celery = "^5.5.3"
kombu = "^5.5.4"
redis = "^5.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^0.24.0"
httpx = "^0.28.1"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.17.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "uvicorn app.main:app --reload"
start-prod = "uvicorn app.main:app --host 0.0.0.0 --port 8000"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true